much clean up of module paths after files were moved.
added an ignore_files option to testing so that you could
omit the testing of specific modules.  This is necessary so that
we can omit modules that will break the test process if they are
imported -- such as plt importing wxPython on a platform without
GTK support (it panics and quits).


git-svn-id: http://svn.scipy.org/svn/scipy/trunk@91 d6536bca-fef9-0310-8506-e4c0a848fbcf