Using get_info('x11') for X11 related include/library files. This fixes 'cannot find -lX11' errors.
Note that much of the contents of the method config_pygist.config_x11
is now disabled as it appears to be an overkill: IMHO, setup_xxx.py
file should not contain sanity tests of external libraries or
the corresponding code should be in some pyautoconf module, at least.
I think, for Scipy we can assume that external libraries are installed
and work properly, otherwise maintaining Scipy becomes very difficult.


git-svn-id: http://svn.scipy.org/svn/scipy/trunk@698 d6536bca-fef9-0310-8506-e4c0a848fbcf