An initial effort at merging PySparse into scipy.  The basic spmatrixmodule.c file now compiles on my machine (x86 Linux) against the array objects in the new scipy core.  But it's still totally untested.
I've added this to the README file:

This is the beginning of a merge of Roman Geus's PySparse into scipy.
The goals are to make it more 'Pythonic', integrate sparse matrix types
into scipy so they act similarly to dense arrays, and to support more
a few more sparse data types, particularly CSC, in a nice OO hierarchy.





git-svn-id: http://svn.scipy.org/svn/scipy/branches/newscipy@1359 d6536bca-fef9-0310-8506-e4c0a848fbcf