Fix #640: bad output from scipy.special.iv(k,z) for large real z.
This occurred because Cephes hyperg.c:hyp1f1p estimated the error
for power series expansion of Kummer confluent 1F1 as zero when
the number of terms was too large. This commit makes Cephes to
assume 100% error when the series does not converge, which makes
it to switch to an asymptotic expansion.

The Cephes hyp1f1 is not exposed by scipy.special (it uses Specfun's
implementation), but it's used internally by Cephes.

Improved tests for Iv are included.

git-svn-id: http://svn.scipy.org/svn/scipy/trunk@5477 d6536bca-fef9-0310-8506-e4c0a848fbcf