Fixing #503: reimplement cephes/iv
Cephes's implementation of Iv via a relation to hyp1f1 is problematic,
especially in the asymptotic regime where intermediate results overflow.

This commit introduces a new implementation for the Bessel I function,
partly converted from the one in the Boost C++ library:

- For large orders, the uniform asymptotic expansion is used.

- Otherwise, use Temme's method.

Performance-wise, this implementation seems to be typically about 2-4
times faster than the Amos routines.

git-svn-id: http://svn.scipy.org/svn/scipy/trunk@5640 d6536bca-fef9-0310-8506-e4c0a848fbcf