Workaround for PyInt_Check failures for python 2.6
If an int is an instance of np.int32, PyInt_Check on it will fail for
python 2.6 and later. Instead of converting the related argument to a
'pure' python int in the Data class, I added a workaround at the C level
to avoid changing anything in the ODR API by accident.

git-svn-id: http://svn.scipy.org/svn/scipy/trunk@5796 d6536bca-fef9-0310-8506-e4c0a848fbcf