Single-tree algorithm to extract pairs.
Users requested a kd-tree routine to extract all distinct pairs of
points within some distance of each other. The existing algorithm
requires two trees and returns all pairs of an element with itself,
while a single-tree implementation could be faster by: not returning
"pairs" with only one element, and by trimming the tree traversal so
that each pair of subtrees is traversed only once. This code does that,
though not in an optimal way.

git-svn-id: http://svn.scipy.org/svn/scipy/trunk@6098 d6536bca-fef9-0310-8506-e4c0a848fbcf