ENH: new faster implementation of stats.kendalltau.
The new tests pass with both the old and new implementations. They do not agree
with the version in mstats though - the tie handling there seems different. For
(long) discussion on kendalltau, see scipy-dev around Mar 2009.

Thanks to Enzo Michelangeli for original implementation, David Simcha and
Sturla Molden for alternate implementations, and Josef for extensive testing.

This closes tickets #999 and #893.

git-svn-id: http://svn.scipy.org/svn/scipy/trunk@6957 d6536bca-fef9-0310-8506-e4c0a848fbcf